additional cleanup; bypass support
